ToroRIL: Recover data connection after leaving airplane mode.
There has been a bug in all CyanogenMod ROMs for Toro since 4.3 where a data connection
can not be established after leaving airplane mode. To get it back we must first kill the
phone process. This is obviously not the most elegant solution, but it works for the
time being.