Mute: Restore Mute functionality through RIL for legacy devices [1/2]
This restores the old Mute function as well as the logic and status bar
notification for devices that do not support muting the microphone through
the system audio manager.

When a call is in progress, start a Media Button callback which will mute
and unmute the microphone using the HEADSETHOOK Keyevent. The statusbar
notification has been moved to this app because Telecom doesn't know the mute
state of the microphone.

We will change Telecom to disable its Media Button callback, and send a HEADSETHOOK
Keyevent when the in-call mute button is pressed.

Issues with this method are that the in-call microphone button never gets
highlighted when we are in mute - but the statusbar icon will indicate whether
we are in a mute state.

add new build.prop entry to enable this: ro.telephony.legacy_mute=1

(See Telecomm for part 2/2)

Good enough for now....